native: macos: add function to get usb ports path
Function builds the usb ports path based on the LocationID property. The same algorithm PySerial uses: https://github.com/pyserial/pyserial/blob/master/serial/tools/list_ports_osx.py#L218

LocationID is used to uniquely identify a device or interface and it's supposed to remain constant across reboots as long as the USB topology doesn't change. It is a 32-bit word.  The top 2 nibbles (bits 31:24) represent the USB Bus Number. Each nibble after that (e.g. bits 23:20 or 19:16) correspond to the port number of the hub that the device is connected to.

IOKit IORegistryEntryGetLocationInPlane represents LocationID as a hex integer converted to string. For example, "14324000". The first 2 digits are bus number. Other digits are ports numbers. We use bus number and ports in dec format, but IORegistryEntryGetLocationInPlane returns them in hex format. Thus each port and bus number convert to from hex to dec. So, "14324000" LocationID is "20-3.2.4" usbPortsPath.